reportclient: return full path to downloaded debuginfo packages
There is no need to try to reconstruct the full path to the downloaded
package when both yum and dnf already know it. Also unpack_rpm()
function appeared to me a bit overloaded because I think this
function should have just one responsibility and that is to unpack
rpm.

Related to #291

Signed-off-by: Jakub Filak <jfilak@redhat.com>